Node - Move extraction & add comments
