+++ bindings/node/native/src/lib.rs
@@ -10,6 +10,7 @@ extern crate tokenizers as tk;
 mod container;
 mod decoders;
 mod encoding;
+mod extraction;
 mod models;
 mod normalizers;
 mod pre_tokenizers;
